refactor: split runtime registration types
