+++ src/CADShared/Cad/Runtime/Registration/AssemInfo.cs
@@ -43,61 +43,3 @@ public struct AssemInfo
     /// </summary>
     public string Description = "";
 }
-
-/// <summary>
-/// 程序集加载类型
-/// </summary>
-public enum AssemLoadType
-{
-    /// <summary>
-    /// 启动
-    /// </summary>
-    Starting = 2,
-
-    /// <summary>
-    /// 随命令
-    /// </summary>
-    ByCommand = 12,
-
-    /// <summary>
-    /// 无效
-    /// </summary>
-    Disabled = 20
-}
-
-/// <summary>
-/// 注册中心配置信息
-/// </summary>
-[Flags]
-public enum AutoRegConfig
-{
-    /// <summary>
-    /// 不进行任何操作
-    /// </summary>
-    Undefined = 0,
-
-    /// <summary>
-    /// 注册表
-    /// </summary>
-    Regedit = 1,
-
-    /// <summary>
-    /// 反射特性
-    /// </summary>
-    ReflectionAttribute = 2,
-
-    /// <summary>
-    /// 反射接口
-    /// </summary>
-    ReflectionInterface = 4,
-
-    /// <summary>
-    /// 移除教育版
-    /// </summary>
-    RemoveEMR = 8,
-
-    /// <summary>
-    /// 全部
-    /// </summary>
-    All = Regedit | ReflectionAttribute | ReflectionInterface | RemoveEMR,
-}
