+++ src/CADShared/Cad/Runtime/Registration/AssemLoadType.cs
@@ -1,0 +1,22 @@
+namespace Fs.Fox.Cad;
+
+/// <summary>
+/// 程序集加载类型
+/// </summary>
+public enum AssemLoadType
+{
+    /// <summary>
+    /// 启动
+    /// </summary>
+    Starting = 2,
+
+    /// <summary>
+    /// 随命令
+    /// </summary>
+    ByCommand = 12,
+
+    /// <summary>
+    /// 无效
+    /// </summary>
+    Disabled = 20
+}
